extract resolve services, add api-key auth
